requirements.txt
anyio==3.6.2
fastapi==0.85.1
idna==3.4
pydantic==1.10.2
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.4.0
boto3==1.24.95
botocore==1.27.95
s3transfer==0.6.0
uvicorn==0.19.0
python-dotenv==0.21.1